requirements.txt
cognitive_face
pyyaml
opencv-python
pyfirmata
gevent
